Updated and debugged all HTML files. There are still a couple errors on the <input> tags for quantity. Not sure what is causing it, so I left it alone, as they were working. refined the search by sport list to include all sports.
Finally got confirmation page working fully and polished. It also has a print function that will allow it to print the confirmation or save it as a .pdf.

Order ID and Payment ID are now pulled into the session when the database creates a new id using the auto number function. This takes place imediatly after the newly created inputDBPayment() and inputDBORder() execute their sql. It does so by calling the newIDGrab() in each BO.

All .jsp files will now display the customers first name on the my account button instead of the word "my"